Update Nutanix Prism Go Client imports to reflect refactored, shallow paths.
