+++ pkg/types/nutanix/client.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"time"
 
-	nutanixclient "github.com/nutanix-cloud-native/prism-go-client/pkg/nutanix"
-	nutanixclientv3 "github.com/nutanix-cloud-native/prism-go-client/pkg/nutanix/v3"
+	nutanixclient "github.com/nutanix-cloud-native/prism-go-client"
+	nutanixclientv3 "github.com/nutanix-cloud-native/prism-go-client/v3"
 )
 
 // CreateNutanixClient creates a Nutanix V3 Client
